scripts/requirements.txt
CairoSVG==2.4.2
Pillow==7.1.2
python-frontmatter==0.5.0